requirement.txt
django==3.0.6
djangorestframework
markdown
psycopg2